refactor: split secret scanning responsibilities
